use map and dynamic lookup but will need k8s client
